refactor: rename human tool marker file
